Add files via upload
These are the audio files for the animal narratives. We stored them on an Micro SD-Card that is inserted in the Catalex Serial MP3-Player. Each folder is for one specific species containing their vocalizations of the green, yellow and red state. The folders are sorted in alphabetical order, starting with bumblebee and ending with polarbear. Folder 8 contains the audio of a species being extinct. The use of the library demanded saving it with the names and the structure as it is now.